Finished homework 2 and made changes from int type input to float

+++ practice_2/cosh_program.py
@@ -1,5 +1,5 @@
 import math
-x = int(input("Input x:"))
+x = float(input("Input x:"))
 cosh = math.cosh(x)
 exp = math.exp(x)
 e = math.e
